feat: Add passwordSharingPaywall
