+++ react/Paywall/PasswordSharingPaywall.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import Paywall from './Paywall'
+
+const PasswordSharingPaywall = ({ onClose }) => {
+  return (
+    <Paywall variant="passwordSharing" onClose={onClose} isPublic={false} />
+  )
+}
+
+export default PasswordSharingPaywall
