+++ react/Paywall/index.jsx
@@ -1,1 +1,2 @@
 export { default as OnlyOfficePaywall } from './OnlyOfficePaywall'
+export { default as PasswordSharingPaywall } from './PasswordSharingPaywall'
